Weak tools and soul stone updates

+++ src/main/java/team/_0mods/ecr/common/utils/MRUWeapon.kt
@@ -1,5 +1,5 @@
 package team._0mods.ecr.common.utils
 
 interface MRUWeapon {
-    val multiplier: Int
+    val multiplier: Float
 }
